Accusation work part 1
On server added a check to verify if accusation correct. If correct, now players notified  with results and message that game is over. If accusation not correct, we still need to do work inactivating the player. Added TODO there.

On client, implemented case for CMD_GAME_OVER and added needed slots/signals for gameover support.

Thank you George for all your help here.